photox_backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
python-dotenv==0.21.1
mysqlclient==2.2.0
gunicorn==23.0.0
django-cors-headers==4.3.1
Pillow==11.2.1
qiniu==7.9.0
requests==2.32.3
torch==2.0.1
torchvision==0.15.2
sympy==1.12
colorthief==0.2.1
django-storages==1.14.2
scikit-image==0.25.2
opencv-contrib-python==4.11.0.86
